# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Create an account and obtain an access + refresh token pair",
        "description": "On success, the refresh token is BOTH set as an `httpOnly` cookie (for web clients) AND returned in the JSON body (for mobile / non-cookie clients).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created and authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthResponseDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Authenticate and receive an access + refresh token pair",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthResponseDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Rotate a refresh token into a new access + refresh pair",
        "description": "The refresh token is read from the `refresh_token` httpOnly cookie if present, otherwise from the JSON body. The old token is invalidated; replaying a revoked token revokes the entire user refresh chain.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthResponseDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh token is missing, invalid, expired, or reused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Revoke a single refresh token (sign-out for this device)",
        "description": "Clears the `refresh_token` cookie and deletes the matching row. Idempotent — succeeds even if the token is already unknown.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh token revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "example": {
                        "message": "Logged out"
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/learning/daily": {
      "get": {
        "operationId": "LearningController_getDailyWords",
        "summary": "Get today's words to review (spaced repetition)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DailyWordResponseDto"
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Learning"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/learning/review": {
      "post": {
        "operationId": "LearningController_reviewWord",
        "summary": "Submit a review for a word",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewWordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReviewResultDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "UserWord does not exist or does not belong to user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Learning"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EnglishFlow API",
    "description": "REST API for the EnglishFlow vocabulary learning app. All successful responses are wrapped in `{ success, data, timestamp }` and all errors in `{ success: false, statusCode, message, error, errors?, path, timestamp }`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "8c5f4e64-1c39-4a55-9c8e-1e1b9c41a0a2",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "student@example.com",
            "format": "email"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-05-11T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-05-11T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "Short-lived JWT (default 15 min). Send as `Authorization: Bearer <token>` on authenticated requests."
          },
          "refreshToken": {
            "type": "string",
            "example": "h6XdR3pK8c-NlYvQ3jKz2hYW9bP5MZeT7vL_8nXrSc",
            "description": "Opaque long-lived refresh token (default 30 days). POST to /auth/refresh to rotate this into a new access + refresh pair when the access token expires."
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "student@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!",
            "minLength": 6,
            "description": "Plain text password, hashed server-side with bcrypt"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "errors": {
            "example": [
              "email must be an email"
            ],
            "description": "Present for validation errors",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/auth/register"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-11T12:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "error",
          "path",
          "timestamp"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "student@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Opaque refresh token. Optional when the request carries the `refresh_token` httpOnly cookie (web clients).",
            "example": "h6XdR3pK8c-NlYvQ3jKz2hYW9bP5MZeT7vL_8nXrSc"
          }
        }
      },
...
      "DailyWordResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UserWord id (use in /learning/review)"
          },
          "wordId": {
            "type": "string",
            "format": "uuid"
          },
          "word": {
            "type": "string",
            "example": "serendipity"
          },
          "translation": {
            "type": "string",
            "example": "kutilmagan yoqimli kashfiyot"
          },
          "example": {
            "type": "object"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "LEARNING",
              "LEARNED"
            ],
            "example": "NEW"
          },
          "repetitionCount": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "id",
          "wordId",
          "word",
          "translation",
          "status",
          "repetitionCount"
        ]
      },
      "ReviewResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "word": {
            "type": "string",
            "example": "serendipity"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "LEARNING",
              "LEARNED"
            ],
            "example": "LEARNING"
          },
          "repetitionCount": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "id",
          "word",
          "status",
          "repetitionCount"
        ]
      },
      "ReviewWordDto": {
        "type": "object",
        "properties": {
          "userWordId": {
            "type": "string",
            "example": "7a1e0b8f-9c12-4d3a-b8e7-65b1d3c9c2f4",
            "format": "uuid",
            "description": "ID of the UserWord (not the Word itself)"
          },
          "correct": {
            "type": "boolean",
            "example": true,
            "description": "Whether the user answered correctly"
          }
        },
        "required": [
          "userWordId",
          "correct"
        ]
      },
...
    }
  }
}